Split admindocs regex utilities into a separate module
